Tidy core types and remove unused calibration lookup
Type band_cwl as list[float] to match its float usage across the reader
and Band.cwl_nm. Remove the unused CalibrationParameters.for_band(); the
(band, start_row, tdi) lookup is the authoritative selector.

+++ src/spp/core/acquisition.py
@@ -40,7 +40,7 @@ class ImagerConfiguration:
     spectral_bands: int
     band_setup: list[int]
     band_start_row: list[int]
-    band_cwl: list[int]
+    band_cwl: list[float]
     scan_direction: int
     binning_factor: int
